US5378 renaming classes to models and setting stripe key in one place
