+++ src/app/billing/billing.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { CreditCardValidator } from '../validators/credit-card.validator';
-import { CustomerBank } from '../classes/customer-bank';
-import { CustomerCard} from '../classes/customer-card';
+import { CustomerBank } from '../models/customer-bank';
+import { CustomerCard} from '../models/customer-card';
 import { GiftService } from '../services/gift.service';
 import { StripeService } from '../services/stripe.service';
 import { PaymentService } from '../services/payment.service';
