+++ src/app/donation/donation.component.ts
@@ -11,7 +11,6 @@ export class DonationComponent implements OnInit {
   constructor(public gift: GiftService) {}
 
   ngOnInit() {
-    (<any>window).Stripe.setPublishableKey(process.env.STRIPE_PUBKEY);
   }
 
 }
