+++ src/app/models/customer-card.ts
