+++ src/app/payment/payment.component.ts
@@ -22,10 +22,7 @@ export class PaymentComponent implements OnInit {
               private fb: FormBuilder) {
   }
 
-  ngOnInit() {
-
-    (<any>window).Stripe.setPublishableKey(process.env.STRIPE_PUBKEY);
-
+  ngOnInit() {        
     if (this.gift.type === 'donation') {
       this.router.navigateByUrl('/donation');
     }
